+++ seed/static/seed/partials/api_docs.html
@@ -1,1 +1,1 @@
-<iframe id="swagger-frame" src="/app/api/swagger/" frameborder="0"></iframe>
+<iframe id="swagger-frame" src="/api/swagger/" frameborder="0"></iframe>
